got session storage up and running
