+++ src/index.html
@@ -20,7 +20,7 @@
     <!--log in/ sign up container-->
     <div id="welcomeFormContainer">
         <h2>Sign Up/Log In To Join Your Troop:</h2>
-    </div>
+
     <form id="userId" action="">
         <h3>Email:</h3>
         <input type="userEmail" name="userEmail" id="userEmailId"></h2>
@@ -30,9 +30,10 @@
         <input type="confirmPass" name="confirmPass" id="confirmPassId"></h2>
         <div>
             <button type="button" class="btn" id="addUserButton">
-            <p>Sign Up</p>
+            Sign Up
          </button>
     </form>
+</div>
     </div>
     </div>
 
@@ -90,32 +91,31 @@
         </div>
         <div class="form-group">
             <label for="eventDescription">Description</label>
-            <textarea id=“eventDescriptionId” rows=“10 " cols=“40”name=“event description”></textarea>
+            <textarea id=“eventDescriptionId” rows="10" cols=“40” name="event" description”></textarea>
         </div>
-        <button type="submit " class="btn btn-primary ">Submit</button>
+        <button type="submit" class="btn btn-primary">Submit</button>
     </form>
 
 
-    <section class="userList "></section>
-    <button id="showUsers " class="button ">butt</button>
+    <section class="userList"></section>
 </body>
 <footer>
     <p>&copy; TFBDP</p>
-    <ul class="footer-links ">
+    <ul class="footer-links">
         <li>
-            <a href="https://www.scouting.org/ "><img class="baby " src="src/images/boyscoutslogo.jpg " alt="Boy Scouts Homepage " style="width:42px;height:42px;border:0; "></a>
+            <a href="https://www.scouting.org/ "><img class="baby" src="src/images/boyscoutslogo.jpg" alt="Boy Scouts Homepage" style="width:42px;height:42px;border:0;"></a>
         </li>
         <li>
-            <a href="https://www.instagram.com/boyscoutsofamerica/?hl=en "><img class="baby " src="src/images/instagram.png " alt="Boy Scouts Instagram " style="width:42px;height:42px;border:0; "></a>
+            <a href="https://www.instagram.com/boyscoutsofamerica/?hl=en"><img class="baby" src="src/images/instagram.png" alt="Boy Scouts Instagram" style="width:42px;height:42px;border:0; "></a>
         </li>
         <li>
-            <a href="https://www.youtube.com/channel/UCbFEBamXVItp6yWqzJvfYFg "><img class="baby " src="src/images/youtube.png " alt="Boy Scouts Youtube channel " style="width:42px;height:42px;border:0; "></a>
+            <a href="https://www.youtube.com/channel/UCbFEBamXVItp6yWqzJvfYFg "><img class="baby" src="src/images/youtube.png" alt="Boy Scouts Youtube channel" style="width:42px;height:42px;border:0; "></a>
         </li>
         <li>
-            <a href="https://twitter.com/boyscouts?lang=en "><img class="baby " src="src/images/twitter.png " alt="Boy Scouts Twitter " style="width:42px;height:42px;border:0; "></a>
+            <a href="https://twitter.com/boyscouts?lang=en "><img class="baby " src="src/images/twitter.png" alt="Boy Scouts Twitter" style="width:42px;height:42px;border:0;"></a>
         </li>
         <li>
-            <a href="https://www.facebook.com/theboyscoutsofamerica/ "><img class="baby " src="src/images/facebook.png " alt="Boy Scouts Facebook " style="width:42px;height:42px;border:0; "></a>
+            <a href="https://www.facebook.com/theboyscoutsofamerica/ "><img class="baby " src="src/images/facebook.png " alt="Boy Scouts Facebook " style="width:42px;height:42px;border:0;"></a>
         </li>
         <li>
             <a href="https://github.com/nss-day-cohort-35/Troop-Fire-Breathing-Duck-Patrol "><img class="baby " src="src/images/firebreathingduck.png " alt="github logo " style="width:42px;height:42px;border:0; "></a>
